+++ lib/src/widgets/raw_editor/raw_editor_state_text_input_client_mixin.dart
@@ -10,6 +10,7 @@ import 'package:flutter/services.dart';
 import '../../models/documents/attribute.dart';
 import '../../models/documents/document.dart';
 import '../../utils/delta.dart';
+import '../../utils/font.dart';
 import '../editor/editor.dart';
 import 'raw_editor.dart';
 
@@ -213,6 +214,20 @@ mixin RawEditorStateTextInputClientMixin on EditorState
           Attribute.fromKeyValue(
             Attribute.font.key,
             widget.configurations.controller.selectedFontFamily,
+          ),
+        );
+      }
+
+      if (widget.configurations.controller.selectedFontSize != null) {
+        widget.configurations.controller.formatText(
+          diff.start,
+          diff.deleted.length,
+          Attribute.fromKeyValue(
+            Attribute.size.key,
+            widget.configurations.controller.selectedFontSize == '0'
+                ? null
+                : getFontSize(
+                    widget.configurations.controller.selectedFontSize),
           ),
         );
       }