+++ packages/quill_html_converter/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## 9.0.0-dev-9
+* Improves the new logic of pasting HTML contents into the Editor
+* Update `README.md` and the doc
+* Dispose the `QuillToolbarSelectHeaderStyleButton` state listener in `dispose`
+* Upgrade the font family button to material 3
+* Rework the font family and font size functionallities to change the font once and type all over the editor
+
 ## 9.0.0-dev-8
 * Better support for pasting HTML contents from external websites to the editor
 * The experimental support of converting the HTML from `quill_html_converter` is now built-in in the `flutter_quill` and removed from there (Breaking change for `quill_html_converter`)